refactor: avoid clangd warnings

+++ src/video/openh264_decoder.cpp
@@ -120,9 +120,11 @@ void OpenH264Decoder::updateImageByTimestamp(const std::uint64_t timestamp) {
           m_tmp_yuv, &buffer_info);
       if (ret != 0) {
         spdlog::error(
-            "OpenH264Decoder DecodeFrameNoDelay failed: error_code={}", ret);
-        throw std::runtime_error(fmt::format(
-            "m_decoder->DecodeFrameNoDelay() failed: error_code={}", ret));
+            "OpenH264Decoder DecodeFrameNoDelay failed: error_code={}",
+            static_cast<std::uint32_t>(ret));
+        throw std::runtime_error(
+            fmt::format("m_decoder->DecodeFrameNoDelay() failed: error_code={}",
+                        static_cast<std::uint32_t>(ret)));
       }
       m_next_timestamp = static_cast<std::uint64_t>(m_webm->getTimestamp());
       if (buffer_info.iBufferStatus == 1) {
